vercel fan/api/requirements.txt
requests
bcrypt
PyJWT